Move CitiesSuggestionBar and SearchBar to `components` folder
